Do not use typedef for the USB Setup struct
Using typedef for structs was relevant in C code to be able to leave out
"struct" when using the type. In C++, this is no longer needed. Removing
the typedef makes it easier to forward-declare the Setup struct later.

Since USBAPI.h can only be included from C++ code, removing this typedef
should not break any code.